backend/requirements.txt
fastapi
uvicorn
yfinance
chromadb
sentence-transformers
groq
python-dotenv
requests
beautifulsoup4
